+++ src/components/analyzers/audioFFTAnalyzer.tsx
@@ -8,6 +8,7 @@ import {
 import MicrophoneAudioControls from "../audio/sourceControls/mic";
 import FFTAnalyzerControls from "./fftAnalyzerControls";
 import FFTAnalyzer from "./analyzers/fft";
+import { useMicrophoneLink } from "./analyzers/common";
 
 interface InternalAudioAnalyzerProps {
   audioSource: AudioSource;
@@ -53,24 +54,6 @@ const InternalAudioFFTAnalyzer = ({
     </>
   );
 };
-
-function useMicrophoneLink(audio: HTMLAudioElement, analyzer: FFTAnalyzer) {
-  return {
-    onMicDisabled: () => {
-      analyzer.disconnectInputs();
-    },
-    onStreamCreated: (stream: MediaStream) => {
-      // Disable any audio
-      audio.pause();
-      // create stream using audio context
-      const streamSrc = analyzer._audioCtx.createMediaStreamSource(stream);
-      // connect microphone stream to analyzer
-      analyzer.connectInput(streamSrc);
-      // mute output to prevent feedback loops from the speakers
-      analyzer.volume = 0.0;
-    },
-  };
-}
 
 interface InternalMicrophoneFFTAnalyzerProps {}
 const InternalMicrophoneFFTAnalyzer =
